# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "algo-coach",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/cards/{slug}": {
      "get": {
        "summary": "Card",
        "operationId": "card_api_cards__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Studied"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/{slug}/recall": {
      "get": {
        "summary": "Prompt",
        "operationId": "prompt_api_cards__slug__recall_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prompted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/{slug}/recall/{template_id}/hints/{hint}": {
      "get": {
        "summary": "Hinted",
        "operationId": "hinted_api_cards__slug__recall__template_id__hints__hint__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "hint",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Hint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hinted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/techniques/{technique}/cards": {
      "get": {
        "summary": "Cards",
        "operationId": "cards_api_techniques__technique__cards_get",
        "parameters": [
          {
            "name": "technique",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Technique"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Card"
                  },
                  "title": "Response Cards Api Techniques  Technique  Cards Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/techniques/{technique}/candidates": {
      "get": {
        "summary": "Offered",
        "operationId": "offered_api_techniques__technique__candidates_get",
        "parameters": [
          {
            "name": "technique",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Technique"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Candidate"
                  },
                  "title": "Response Offered Api Techniques  Technique  Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/problems/{problem_id}": {
      "get": {
        "summary": "Picked",
        "operationId": "picked_api_problems__problem_id__get",
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Problem Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Picked"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/problems/{problem_id}/sittings": {
      "post": {
        "summary": "Statement",
        "operationId": "statement_api_problems__problem_id__sittings_post",
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Problem Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Served"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sittings/{sitting_id}": {
      "get": {
        "summary": "Sitting",
        "operationId": "sitting_api_sittings__sitting_id__get",
        "parameters": [
          {
            "name": "sitting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sitting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Served"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Candidate": {
        "properties": {
          "problem_id": {
            "type": "string",
            "title": "Problem Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "difficulty": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemDifficulty"
              },
              {
                "type": "null"
              }
            ]
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "solved_count": {
            "type": "integer",
            "title": "Solved Count"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          }
        },
        "type": "object",
        "required": [
          "problem_id",
          "title",
          "difficulty",
          "attempt_count",
          "solved_count",
          "last_attempt_at"
        ],
        "title": "Candidate",
        "description": "A problem offered for a drill. The statement is left out: the clock\nstarts when it is served."
      },
      "Card": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]*$",
            "title": "Slug"
          },
          "technique": {
            "type": "string",
            "title": "Technique"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "trigger": {
            "type": "string",
            "minLength": 1,
            "title": "Trigger"
          },
          "brief": {
            "type": "string",
            "minLength": 1,
            "title": "Brief"
          },
          "templates": {
            "items": {
              "$ref": "#/components/schemas/Template"
            },
            "type": "array",
            "minItems": 1,
            "title": "Templates"
          },
          "selector": {
            "$ref": "#/components/schemas/Selector"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "technique",
          "title",
          "trigger",
          "brief",
          "templates",
          "selector"
        ],
        "title": "Card",
        "description": "Teaching content for one technique: forms to reproduce, and a\nselector."
      },
      "CardRun": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "minLength": 1,
            "title": "User Id"
          },
          "card_id": {
            "type": "string",
            "minLength": 1,
            "title": "Card Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "probes": {
            "items": {
              "$ref": "#/components/schemas/Probe"
            },
            "type": "array",
            "title": "Probes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "card_id",
          "started_at"
        ],
        "title": "CardRun"
      },
...
      "Gap": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "template_id",
          "slug",
          "title"
        ],
        "title": "Gap",
        "description": "A core template no solution displays, named as the card names it.\n\nThe card claims to teach the form, so the ladder reports the gap rather\nthan substituting another problem for the rung. The next generation run is\naimed at what this names."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Hint": {
        "type": "string",
        "enum": [
          "title",
          "notes",
          "form"
        ],
        "title": "Hint",
        "description": "What the trainer gives when memory fails, in the order it gives them.\nEach answers more of the question than the last."
      },
      "Hinted": {
        "properties": {
          "hint": {
            "$ref": "#/components/schemas/Hint"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "hint",
          "text"
        ],
        "title": "Hinted"
      },
...
      "Picked": {
        "properties": {
          "problem_id": {
            "type": "string",
            "title": "Problem Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "difficulty": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemDifficulty"
              },
              {
                "type": "null"
              }
            ]
          },
          "techniques": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Techniques"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "solved_count": {
            "type": "integer",
            "title": "Solved Count"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          }
        },
        "type": "object",
        "required": [
          "problem_id",
          "title",
          "difficulty",
          "techniques",
          "attempt_count",
          "solved_count",
          "last_attempt_at"
        ],
        "title": "Picked",
        "description": "The problem a user picked, before its statement is served. The statement\nis left out, as a candidate's is."
      },
      "Probe": {
        "properties": {
          "problem_id": {
            "type": "string",
            "minLength": 1,
            "title": "Problem Id"
          },
          "assigned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assigned At"
          }
        },
        "type": "object",
        "required": [
          "problem_id",
          "assigned_at"
        ],
        "title": "Probe",
        "description": "One problem a run was given, testing whether the form is recognised\nunprompted. Drawn from outside the ladder, which teaches the form."
      },
      "Probed": {
        "properties": {
          "problem": {
            "$ref": "#/components/schemas/Problem"
          },
          "assigned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assigned At"
          },
          "attempted": {
            "type": "boolean",
            "title": "Attempted"
          }
        },
        "type": "object",
        "required": [
          "problem",
          "assigned_at",
          "attempted"
        ],
        "title": "Probed",
        "description": "One probe the start drew, named as the page shows it. A probe tests\nwhether the form is recognised unprompted, so the page names the problem\nand says nothing of the technique behind it."
      },
...
      "ProblemDifficulty": {
        "type": "string",
        "enum": [
          "easy",
          "medium",
          "hard"
        ],
        "title": "ProblemDifficulty"
      },
...
      "Prompted": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "trigger": {
            "type": "string",
            "title": "Trigger"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          }
        },
        "type": "object",
        "required": [
          "template_id",
          "trigger",
          "signature"
        ],
        "title": "Prompted",
        "description": "What the trainer shows: the template's trigger and the signature its\ncases call. The title and the form are withheld, and mapping the trigger\nto the form is the recall being measured."
      },
...
      "Recalled": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "last_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last At"
          },
          "hints": {
            "items": {
              "$ref": "#/components/schemas/Hint"
            },
            "type": "array",
            "title": "Hints"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "template_id",
          "last_at",
          "hints",
          "verified"
        ],
        "title": "Recalled",
        "description": "One template's recall state: the last reproduction of it, and nothing\nof the ones before, which stay in the log."
      },
...
      "Rung": {
        "properties": {
          "problem": {
            "$ref": "#/components/schemas/Problem"
          },
          "templates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Templates"
          },
          "required": {
            "type": "boolean",
            "title": "Required"
          },
          "solved": {
            "type": "boolean",
            "title": "Solved",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "problem",
          "templates",
          "required"
        ],
        "title": "Rung",
        "description": "One problem on a ladder, with the templates its canonicals display.\n\nA rung the selector filled covers none. `required` is derived from what the\nrung covers rather than stored, and a rung covering the optional template\nbeside a core one offers that form as the alternative approach."
      },
      "Selector": {
        "properties": {
          "technique": {
            "type": "string",
            "title": "Technique"
          },
          "difficulty": {
            "items": {
              "$ref": "#/components/schemas/ProblemDifficulty"
            },
            "type": "array",
            "title": "Difficulty"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "technique",
          "size"
        ],
        "title": "Selector",
        "description": "What a ladder resolves from. Named fields rather than a filter map, so\na new filter is an additive optional field."
      },
      "Served": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "signature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature"
          },
          "sitting": {
            "$ref": "#/components/schemas/Sitting"
          },
          "elapsed_sec": {
            "type": "number",
            "title": "Elapsed Sec"
          }
        },
        "type": "object",
        "required": [
          "title",
          "statement",
          "signature",
          "sitting",
          "elapsed_sec"
        ],
        "title": "Served",
        "description": "What a solver is handed: the statement and the clock, and nothing the\nproblem was written from."
      },
...
      "Studied": {
        "properties": {
          "card": {
            "$ref": "#/components/schemas/Card"
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CardRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "rungs": {
            "items": {
              "$ref": "#/components/schemas/Rung"
            },
            "type": "array",
            "title": "Rungs"
          },
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/Gap"
            },
            "type": "array",
            "title": "Gaps"
          },
          "recall": {
            "items": {
              "$ref": "#/components/schemas/Recalled"
            },
            "type": "array",
            "title": "Recall"
          },
          "probes": {
            "items": {
              "$ref": "#/components/schemas/Probed"
            },
            "type": "array",
            "title": "Probes"
          }
        },
        "type": "object",
        "required": [
          "card",
          "run",
          "rungs",
          "gaps",
          "recall",
          "probes"
        ],
        "title": "Studied",
        "description": "A card as its page reads it. The ladder, the progress and the recall\nstate are folds, and `content.md` gives why none of them is stored."
      },
...
      "Template": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]*$",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "trigger": {
            "type": "string",
            "minLength": 1,
            "title": "Trigger"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "optional": {
            "type": "boolean",
            "title": "Optional",
            "default": false
          },
          "speedup": {
            "type": "boolean",
            "title": "Speedup",
            "default": true
          },
          "kind": {
            "$ref": "#/components/schemas/TemplateKind",
            "default": "code"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "unordered": {
            "type": "boolean",
            "title": "Unordered",
            "default": false
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/TemplateCase"
            },
            "type": "array",
            "title": "Cases"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "trigger",
          "code"
        ],
        "title": "Template",
        "description": "One form reproduced from memory. The unit of recall."
      },
      "TemplateCase": {
        "properties": {
          "args": {
            "items": {},
            "type": "array",
            "title": "Args"
          },
          "expected": {
            "title": "Expected"
          }
        },
        "type": "object",
        "required": [
          "expected"
        ],
        "title": "TemplateCase",
        "description": "One case a reproduced form is checked against. The arguments are\npositional, and `solve` is the function they call."
      },
      "TemplateKind": {
        "type": "string",
        "enum": [
          "code",
          "procedure"
        ],
        "title": "TemplateKind"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}